requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
mysql-connector-python==8.1.0
mysqlclient==2.2.7
Werkzeug==2.3.7
python-dotenv==1.0.0
email-validator==2.0.0
gunicorn==21.2.0 